+++ yash-syntax/src/parser/command.rs
@@ -22,6 +22,9 @@
 use super::core::Parser;
 use super::core::Rec;
 use super::core::Result;
+use super::lex::Keyword::{Function, OpenBracketBracket};
+use super::lex::TokenId::Token;
+use super::{Error, SyntaxError};
 use crate::syntax::Command;
 
 impl Parser<'_, '_> {
@@ -32,10 +35,28 @@ impl Parser<'_, '_> {
     pub async fn command(&mut self) -> Result<Rec<Option<Command>>> {
         match self.simple_command().await? {
             Rec::AliasSubstituted => Ok(Rec::AliasSubstituted),
-            Rec::Parsed(None) => self
-                .full_compound_command()
-                .await
-                .map(|c| Rec::Parsed(c.map(Command::Compound))),
+
+            Rec::Parsed(None) => {
+                if let Some(compound) = self.full_compound_command().await? {
+                    return Ok(Rec::Parsed(Some(Command::Compound(compound))));
+                }
+
+                let next = self.peek_token().await?;
+                match next.id {
+                    Token(Some(Function)) => {
+                        let cause = SyntaxError::UnsupportedFunctionDefinitionSyntax.into();
+                        let location = next.word.location.clone();
+                        Err(Error { cause, location })
+                    }
+                    Token(Some(OpenBracketBracket)) => {
+                        let cause = SyntaxError::UnsupportedDoubleBracketCommand.into();
+                        let location = next.word.location.clone();
+                        Err(Error { cause, location })
+                    }
+                    _ => Ok(Rec::Parsed(None)),
+                }
+            }
+
             Rec::Parsed(Some(c)) => self
                 .short_function_definition(c)
                 .await
@@ -46,11 +67,13 @@ impl Parser<'_, '_> {
 
 #[cfg(test)]
 mod tests {
+    use super::super::ErrorCause;
     use super::super::lex::Lexer;
     use super::super::lex::TokenId::EndOfInput;
     use super::*;
+    use crate::source::Source;
     use assert_matches::assert_matches;
-    use futures_util::FutureExt;
+    use futures_util::FutureExt as _;
 
     #[test]
     fn parser_command_simple() {
@@ -83,7 +106,7 @@ mod tests {
     }
 
     #[test]
-    fn parser_command_function() {
+    fn parser_command_short_function() {
         let mut lexer = Lexer::with_code("fun () ( echo )");
         let mut parser = Parser::new(&mut lexer);
 
@@ -95,6 +118,38 @@ mod tests {
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
         assert_eq!(next.id, EndOfInput);
+    }
+
+    #[test]
+    fn parser_command_long_function() {
+        let mut lexer = Lexer::with_code("  function fun { echo; }");
+        let mut parser = Parser::new(&mut lexer);
+
+        let e = parser.command().now_or_never().unwrap().unwrap_err();
+        assert_eq!(
+            e.cause,
+            ErrorCause::Syntax(SyntaxError::UnsupportedFunctionDefinitionSyntax)
+        );
+        assert_eq!(*e.location.code.value.borrow(), "  function fun { echo; }");
+        assert_eq!(e.location.code.start_line_number.get(), 1);
+        assert_eq!(*e.location.code.source, Source::Unknown);
+        assert_eq!(e.location.range, 2..10);
+    }
+
+    #[test]
+    fn parser_command_double_bracket() {
+        let mut lexer = Lexer::with_code(" [[ foo ]]");
+        let mut parser = Parser::new(&mut lexer);
+
+        let e = parser.command().now_or_never().unwrap().unwrap_err();
+        assert_eq!(
+            e.cause,
+            ErrorCause::Syntax(SyntaxError::UnsupportedDoubleBracketCommand)
+        );
+        assert_eq!(*e.location.code.value.borrow(), " [[ foo ]]");
+        assert_eq!(e.location.code.start_line_number.get(), 1);
+        assert_eq!(*e.location.code.source, Source::Unknown);
+        assert_eq!(e.location.range, 1..3);
     }
 
     #[test]