+++ yash-prompt/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - External dependency versions:
-    - yash-syntax 0.15.0 → 0.15.1
+    - yash-syntax 0.15.0 → 0.15.2
 - Internal dependency versions:
     - yash-semantics 0.8.0 → 0.9.0